pyproject.toml
[project]
name = "playwright_tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "beautifulsoup4==4.13.4",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "filelock==3.18.0",
    "fsspec==2025.5.1",
    "google-auth==2.40.3",
    "google-genai==1.20.0",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.33.0",
    "idna==3.10",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numpy==2.3.0",
    "packaging==25.0",
    "pandas>=2.3.0",
    "pillow==11.2.1",
    "playwright==1.52.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rsa==4.9.1",
    "safetensors==0.5.3",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sympy==1.14.0",
    "tokenizers==0.21.1",
    "torch==2.7.1",
    "torchvision>=0.22.1",
    "tqdm==4.67.1",
    "transformers==4.52.4",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "urllib3==2.4.0",
    "websockets==15.0.1",
]
